+++ test/nebula_cli_test.dart
@@ -19,7 +19,7 @@ void main() {
       await cliTemp.delete(recursive: true);
     });
 
-    test('CA + signing one machine', () async {
+    test('CA + signing in one step', () async {
       final temp = await Directory.systemTemp.createTemp();
       try {
         final cli = NebulaCli(path: cliTemp.path);
@@ -45,6 +45,42 @@ void main() {
           'machine.crt',
           'machine.crt.json',
           'machine.key',
+          'machine.png',
+        });
+      } finally {
+        await temp.delete(recursive: true);
+      }
+    });
+
+    test('CA + keygen + sign', () async {
+      final temp = await Directory.systemTemp.createTemp();
+      try {
+        final cli = NebulaCli(path: cliTemp.path);
+        final caPrefix = p.join(temp.path, 'test-ca');
+        await cli.ca(name: 'Test CA', outputPrefix: caPrefix);
+        final machinePrefix = p.join(temp.path, 'machine');
+        await cli.keygen(outputPrefix: machinePrefix);
+        await cli.sign(
+          caPrefix: caPrefix,
+          ip: '192.168.100.1/24',
+          name: 'machine',
+          groups: ['a', 'b'],
+          outputPrefix: p.join(temp.path, 'machine'),
+        );
+
+        final files = temp.listSync().whereType<File>();
+        final paths = files
+            .map((e) => e.path)
+            .map((e) => p.relative(e, from: temp.path))
+            .toSet();
+        expect(paths, {
+          'test-ca.crt',
+          'test-ca.crt.json',
+          'test-ca.key',
+          'machine.crt',
+          'machine.crt.json',
+          'machine.key',
+          'machine.pub',
           'machine.png',
         });
       } finally {